Started decoupling the public api of the script from argparse.
There was an interview for The Pokemon Company I did a while back. I
showed this to my interviewer as a "passion project". He asked me:

"Have you considered making this a web app?"

The answer is yes, and I think it's a terrible idea. That would entail
me making use of a cloud service, making necessary changes for security
of that service, creating a UI which probably won't even let you
intensify multiple images at once, all for something where you could
just run the dang thing on your desktop.

But there may come a point where I have to do something unecessary to be
impressive to an interviewer, where somehow, someway it matters that I
design this in ways that appear suboptimal to me, but useful to others.

This is not that commit, but it's a stop on the way there.